pyproject.toml
[tool.poetry]
name = "raspberry_pi_fpga_node"
version = "0.1.0"
description = ""
authors = ["Maksprots <makslive@bk.ru>,", "Marlen2204"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.8"
faststream = "^0.5.34"
sqlalchemy = "^2.0.38"
uvicorn = "^0.34.0"
aio-pika = "^9.5.5" # transien faststream
pydantic-settings = "^2.8.1"
aiobotocore = "^2.21.1"
aiofiles = "^24.1.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.9.10"
pylint = "^3.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
